pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["marisiman <email_dari_marisiman@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
bcrypt = "4.1.2"
blinker = "1.8.1"
click = "8.1.7"
colorama = "0.4.6"
flask = "3.0.3"
flask-bcrypt = "1.0.1"
flask-cors = "4.0.0"
flask-jwt-extended = "4.6.0"
flask-sqlalchemy = "3.1.1"
greenlet = "3.0.3"
gunicorn = "22.0.0"
itsdangerous = "2.2.0"
jinja2 = "3.1.3"
markupsafe = "2.1.5"
packaging = "24.0"
pyjwt = "2.8.0"
pymysql = "1.1.0"
python-dotenv = "1.0.1"
sqlalchemy = "2.0.29"
typing-extensions = "4.11.0"
werkzeug = "3.0.2"
pydantic = "^2.7.1"
flask-login = "^0.6.3"
mysql-connector-python = "^8.4.0"
flask-migrate = "^4.0.7"
waitress = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
